app/requirements.txt
pyjwt
pyjwt[crypto]
flask
flask_sqlalchemy
flask_bcrypt
